StepperTitle as component
